% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mfdp_single.R
\name{read_mfdp_input_table}
\alias{read_mfdp_input_table}
\title{Read single-file MFDP input table format}
\usage{
read_mfdp_input_table(inputfile)
}
\arguments{
\item{inputfile}{A path to the text file which contains the MFDP input table}
}
\value{
An FLStock object with all the data from the MFDP input table
}
\description{
Read single-file MFDP input table format
}
