% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdp.R
\name{mfdp}
\alias{mfdp}
\title{Run MFDP}
\usage{
mfdp(input_file, run_name = "mfdp")
}
\arguments{
\item{input_file}{A path to the MFDP index file}

\item{run_name}{A name for the run. This will be applied in the result file names.}
}
\value{
A list of result objects
}
\description{
Run MFDP
}
