% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdp.R
\name{mfdp}
\alias{mfdp}
\title{Run MFDP}
\usage{
mfdp(input, configs = NULL, run_name = "mfdp", out_dir = tempdir())
}
\arguments{
\item{input}{Input to MFDP. It can be either 1) A path to the MFDP index file, or 2) An FLStock object
from \code{read_mfdp_input_table()}.}

\item{configs}{Path to extra configuration file for MFDP. This may contain TAC,
HCR function, among others.}

\item{run_name}{A name for the run. This will be applied in the result file
names. Default is 'mfdp'.}

\item{out_dir}{Path for the generated output files. Default is \code{tempdir()}.}
}
\value{
A list of result objects
}
\description{
Run MFDP
}
